Moved abstract clients to separated place
