+++ services/assistant/app/infra/grpc/__init__.py
@@ -1,12 +1,10 @@
 from app.infra.grpc.client_factories import XrayClientFactory, BookingClientFactory
-from app.infra.grpc.xray_client import XrayClient, AbstractXrayClient
-from app.infra.grpc.booking_client import BookingClient, AbstractBookingClient
+from app.infra.grpc.xray_client import XrayClient
+from app.infra.grpc.booking_client import BookingClient
 
 __all__ = [
     "BookingClientFactory",
     "XrayClientFactory",
     "XrayClient",
     "BookingClient",
-    "AbstractXrayClient",
-    "AbstractBookingClient",
 ]
