+++ services/assistant/app/infra/grpc/booking_client.py
@@ -1,23 +1,9 @@
-from abc import ABC, abstractmethod
+from app.clients.booking_client import BookingClient as AbstractBookingClient
 
 from app.infra.grpc.gen.booking_pb2_grpc import BookingStub
 from app.infra.grpc.gen.booking_pb2 import BookingRequest, BookingResponse
 
 from app.infra.grpc.grpc_client import GRPCClient
-
-
-class AbstractBookingClient(ABC):
-    @abstractmethod
-    async def run_booking(self, email: str, password: str) -> bool:
-        """:return: `True` if booking ran, `False` otherwise."""
-
-    @abstractmethod
-    async def stop_booking(self, email: str, password: str):
-        pass
-
-    @abstractmethod
-    async def booked(self, email: str, password: str) -> bool:
-        pass
 
 
 class BookingClient(GRPCClient, AbstractBookingClient):
